remove bug from authExceptionHandler
