+++ src/App.js
@@ -1,7 +1,7 @@
 import { HomePage } from "./Components";
 import { Navbar } from "./Components";
 import "./styles.css";
-import { Routes, Route } from "react-router-dom";
+import { Routes, Route, useNavigate } from "react-router-dom";
 import { Playlists } from "./Components";
 import { User } from "./Components";
 import { Watch } from "./Components";
@@ -11,13 +11,31 @@ import { PrivateRoute } from "./Utils/PrivateRoute";
 import { Login } from "./Components/Authentication/Login";
 import { Signup } from "./Components/Authentication/Sigup";
 import { useAuth } from "./Context/UserProvider";
-import { setupAuthHeaderForServiceCalls } from "./Utils/UseAxios";
+import { useEffect } from "react";
+import {
+  setupAuthHeaderForServiceCalls,
+  setupAuthExceptionHandler,
+} from "./Utils/UseAxios";
 
 export default function App() {
-  const { token } = useAuth();
-  if (token) {
-    setupAuthHeaderForServiceCalls(token);
-  }
+  const { token, userDispatch } = useAuth();
+  const navigate = useNavigate();
+
+  const logOutUser = () => {
+    localStorage.removeItem("session");
+    userDispatch({
+      type: "SET_LOGIN",
+      token: null,
+      user: null,
+    });
+  };
+
+  useEffect(() => {
+    if (token) {
+      setupAuthHeaderForServiceCalls(token);
+      setupAuthExceptionHandler(logOutUser, navigate);
+    }
+  }, []);
 
   return (
     <div className="App">
